final push

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: rel-maza <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 18:39:36 by rel-maza          #+#    #+#             */
-/*   Updated: 2021/11/05 11:37:04 by rel-maza         ###   ########.fr       */
+/*   Updated: 2021/11/18 10:10:25 by rel-maza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memset(void *s, int c, size_t n)
 {
-	unsigned int	i;
+	size_t			i;
 	unsigned char	*str;
 
 	str = (unsigned char *) s;
